Use Auth0 Passwordless with the redirect flow
This is basically the only functional option right now.  Auth begins
with a GET to /auth0/login, which shows the passwordless lock (neither
the passwordful lock nor the mozilla lock work, and passwordless lock
doesn't work when embedded in a React application).  On submit, that
redirects to /auth0/callback with the necessary tokens, and that
endpoint translates those into a User object, which is used to generate
credentials as usual (by redirecting to /).