Bug 1385363 - Support converting an auth0 oidc token to TC creds
Given an access_token with an appropriate audience, generate Taskcluster
credentials for the authorized user.

This is a partial implementation: at present, the profiles available
from the auth0 management API are not updated to the new CIS profile, so
this API cannot determine group membership, and thus issues powerless
credentials.