Bug 1395574 - support LDAP logins via auth0
This adds support for logins using an LDAP IdP.  It doesn't change the
UI at all, so users won't use it, but it can be tested by going to
https://login.taskcluster.net/auth0/callback, and we can easily enough
add a link to that in October if necessary.

The identity-matching code is similar to that in
`src/handlers/mozilla-auth0.js`.